Moved util functions to their own files
