+++ .luacheckrc
@@ -1,0 +1,5 @@
+std = "min"
+cache = true
+
+files["spec/"].globals = { "describe", "it" }
+globals = { "describe" }
